# v-cloud-test-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vibbraneo",
        "contact": {
            "email": "diogorcazevedo@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://administracaodosistema.com.br/sv/api"
        },
        {
            "url": "http://127.0.0.1:8000/api/"
        }
    ],
    "paths": {
...
        "/register/show/{id}": {
            "get": {
                "tags": [
                    "register"
                ],
                "summary": "show register ",
                "description": "show register",
                "operationId": "showRegister",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id param ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "descrition list"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/register/store/{id}": {
            "post": {
                "tags": [
                    "register"
                ],
                "summary": "store register",
                "description": "create register",
                "operationId": "registerStore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do usuário requerente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Pass store fields",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Novo Fulano de Tal"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "novofulano@gmail.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "descrition list"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/register/update/{id}": {
            "put": {
                "tags": [
                    "register"
                ],
                "summary": "update register",
                "description": "edit register",
                "operationId": "registerUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": " id param",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Pass Update fields",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Mudei de Nome de Tal"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "mudeideemail@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "descrition list"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/register/updatePassword/{id}": {
            "put": {
                "tags": [
                    "register"
                ],
                "summary": "update Password register",
                "description": "edit Password register",
                "operationId": "registerUpdatePassword",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": " id param",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Pass Update fields",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "password",
                                        "format": "text",
                                        "example": "654321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "descrition list"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/register/updateLocation/{id}": {
            "put": {
                "tags": [
                    "register"
                ],
                "summary": "update Location register",
                "description": "edit Location register",
                "operationId": "registerLocationUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": " id param",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Pass Update fields",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "zip_code",
                                    "state",
                                    "city",
                                    "address"
                                ],
                                "properties": {
                                    "lat": {
                                        "type": "number",
                                        "format": "text",
                                        "example": "-20.311738744829306"
                                    },
                                    "lng": {
                                        "type": "number",
                                        "format": "text",
                                        "example": "-40.288277444412245"
                                    },
                                    "zip_code": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "29055260"
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Rua Aleixo Netto 1226"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Vitória"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "ES"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "descrition list"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}